+++ agent/tests/PendingConnection.php
@@ -21,6 +21,8 @@ class PendingConnection
 
         $connection->emit('end');
 
+        $connection->emit('close');
+
         $connection->removeAllListeners();
         $this->server->connections[] = $connection;
     }