+++ src/Hooks/CommandLifecycleIsLongerThanHandler.php
@@ -26,13 +26,13 @@ final class CommandLifecycleIsLongerThanHandler
     public function __invoke(Carbon $startedAt, InputInterface $input, int $status): void
     {
         try {
-            $this->nightwatch->sensor->stage(ExecutionStage::End);
+            $this->nightwatch->stage(ExecutionStage::End);
         } catch (Throwable $e) {
             $this->nightwatch->report($e);
         }
 
         try {
-            $this->nightwatch->sensor->command($input, $status);
+            $this->nightwatch->command($input, $status);
         } catch (Throwable $e) {
             $this->nightwatch->report($e);
         }