+++ src/NightwatchServiceProvider.php
@@ -67,7 +67,6 @@ use Laravel\Nightwatch\State\RequestState;
 use Throwable;
 
 use function defined;
-use function is_string;
 use function microtime;
 
 /**
@@ -89,6 +88,10 @@ final class NightwatchServiceProvider extends ServiceProvider
     /**
      * @var array{
      *     enabled?: bool,
+     *     sampling: array{
+     *        requests: float,
+     *        commands: float,
+     *     },
      *     token?: string,
      *     deployment?: string,
      *     server?: string,
@@ -182,7 +185,7 @@ final class NightwatchServiceProvider extends ServiceProvider
     {
         $this->app->singleton(RouteMiddleware::class, fn () => new RouteMiddleware($this->core)); // @phpstan-ignore argument.type
 
-        $this->app->scoped(GlobalMiddleware::class, fn () => new GlobalMiddleware($this->core));
+        $this->app->scoped(GlobalMiddleware::class, fn () => new GlobalMiddleware($this->core)); // @phpstan-ignore argument.type
     }
 
     private function registerAgentCommand(): void
@@ -216,7 +219,25 @@ final class NightwatchServiceProvider extends ServiceProvider
             state: $state,
             clock: $clock,
             enabled: ($this->nightwatchConfig['enabled'] ?? true),
+            sampling: [
+                'requests' => $this->configuredSampleRate('requests'),
+                'commands' => $this->configuredSampleRate('commands'),
+            ],
         ));
+    }
+
+    /**
+     * @param  'requests'|'commands'  $key
+     */
+    private function configuredSampleRate($key): float
+    {
+        $value = (float) ($this->nightwatchConfig['sampling'][$key] ?? 1.0);
+
+        if ($value < 0 || $value > 1) {
+            return 0.0;
+        }
+
+        return $value;
     }
 
     private function handleAndClearRegisterException(): void
@@ -335,7 +356,7 @@ final class NightwatchServiceProvider extends ServiceProvider
          *
          * TODO handle this on the queue
          */
-        $events->listen(Logout::class, (new LogoutListener($core))(...)); // @phpstan-ignore argument.type
+        $events->listen(Logout::class, (new LogoutListener($core))(...));
 
         /**
          * @see \Laravel\Nightwatch\ExecutionStage::BeforeMiddleware
@@ -436,17 +457,13 @@ final class NightwatchServiceProvider extends ServiceProvider
             return new CommandState(
                 timestamp: $this->timestamp,
                 trace: new LazyValue(static function () {
-                    $trace = Compatibility::getHiddenContext('nightwatch_trace_id');
+                    return (string) Compatibility::getHiddenContext('nightwatch_trace_id', static function () { // @phpstan-ignore cast.string
+                        $trace = (string) Str::uuid();
 
-                    if (is_string($trace)) {
+                        Compatibility::addHiddenContext('nightwatch_trace_id', $trace);
+
                         return $trace;
-                    }
-
-                    $trace = (string) Str::uuid();
-
-                    Compatibility::addHiddenContext('nightwatch_trace_id', $trace);
-
-                    return $trace;
+                    });
                 }),
                 id: $trace,
                 currentExecutionStageStartedAtMicrotime: $this->timestamp,