+++ CHANGELOG.md
@@ -1,6 +1,37 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/nightwatch/compare/v1.2.1...1.x)
+## [Unreleased](https://github.com/laravel/nightwatch/compare/v1.4.0...1.x)
+
+## [v1.4.0](https://github.com/laravel/nightwatch/compare/v1.3.0...v1.4.0) - 2025-04-28
+
+### What's Changed
+
+* Job sampling by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/147
+* Command sampling by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/148
+* Encapsulate capturing of scheduled tasks by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/150
+* Improve local ci check performance by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/151
+* Fix `facades` build permission by [@crynobone](https://github.com/crynobone) in https://github.com/laravel/nightwatch/pull/149
+
+### New Contributors
+
+* [@crynobone](https://github.com/crynobone) made their first contribution in https://github.com/laravel/nightwatch/pull/149
+
+**Full Changelog**: https://github.com/laravel/nightwatch/compare/v1.3.0...v1.4.0
+
+## [v1.3.0](https://github.com/laravel/nightwatch/compare/v1.2.1...v1.3.0) - 2025-04-24
+
+### What's Changed
+
+* Improve quota check performance by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/136
+* Add status command by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/140
+* Use Bencode for agent payloads  by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/141
+* Fix typo by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/143
+* Request Sampling by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/142
+* Write payload for debugging by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/144
+* Add watch command by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/145
+* Improve status command by [@timacdonald](https://github.com/timacdonald) in https://github.com/laravel/nightwatch/pull/146
+
+**Full Changelog**: https://github.com/laravel/nightwatch/compare/v1.2.1...v1.3.0
 
 ## [v1.2.1](https://github.com/laravel/nightwatch/compare/v1.2.0...v1.2.1) - 2025-04-16